Change multiprocessing on extract
